P-04-02-2025/1

+++ Shared/Models/ElemetZamDoZlecen.cs
@@ -9,13 +9,13 @@ namespace GEORGE.Shared.Models
         [Required]
         public string? RowIdZlecenia { get; set; }
         public string? RowIdProducent { get; set; }
-        public float? Szerokosc { get; set; } = 0;
-        public float? Wysokosc { get; set; } = 0;
-        public float? Dlugosc { get; set; } = 0;
-        public float? Waga { get; set; } = 0;
-        public float? Powierzchnia { get; set; } = 0;
-        public float? Objetosc { get; set; } = 0;
-        public float? CenaNetto { get; set; } = 0;
+        public decimal? Szerokosc { get; set; } = 0;
+        public decimal? Wysokosc { get; set; } = 0;
+        public decimal? Dlugosc { get; set; } = 0;
+        public decimal? Waga { get; set; } = 0;
+        public decimal? Powierzchnia { get; set; } = 0;
+        public decimal? Objetosc { get; set; } = 0;
+        public decimal? CenaNetto { get; set; } = 0;
         public string? Jednostka { get; set; } = "";
         public string? Typ { get; set; } = "";
         public string? NumerKatalogowy { get; set; } = "";
